requirements.txt
einops==0.6.0
numpy==1.23.5
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.10.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1
